NP-1715 Use Formik Field component
